+++ docs/Changelog.md
@@ -8,6 +8,7 @@
 - vtm-mvt module with MVT tile decoder [#481](https://github.com/mapsforge/vtm/pull/481)
 - Nextzen MVT / GeoJSON vector tiles [#498](https://github.com/mapsforge/vtm/issues/498)
 - OpenMapTiles MVT vector tiles [#482](https://github.com/mapsforge/vtm/issues/482)
+- Location texture renderer [#547](https://github.com/mapsforge/vtm/issues/547)
 - Render themes: symbols on lines [#495](https://github.com/mapsforge/vtm/issues/495)
 - Render themes: styles improvements [#479](https://github.com/mapsforge/vtm/pull/479)
 - Internal render themes improvements [#488](https://github.com/mapsforge/vtm/pull/488)